this arrangement seems to work now
Requires the workflow files to be in the package_dir (using symlinks
doesn't work, things end up outside of package dir). Then can create
symlinks in the root directory to enable snakemake from root dir to
work. Perhaps not ideal, but right now it seems this is the only way
to do this with poetry/pyproject (there are a few open issues on
poetry github).